Move mapper from scan result to file record mapper
