+++ apps/server/src/modules/files-storage/mapper/file-record.mapper.ts
@@ -1,5 +1,5 @@
-import { FileRecordListResponse, FileRecordResponse } from '../controller/dto';
-import { FileRecord } from '../entity';
+import { FileRecordListResponse, FileRecordResponse, ScanResultDto, ScanResultParams } from '../controller/dto';
+import { FileRecord, ScanStatus } from '../entity';
 
 export class FileRecordMapper {
 	static mapToFileRecordResponse(fileRecord: FileRecord): FileRecordResponse {
@@ -16,5 +16,25 @@ export class FileRecordMapper {
 
 		const response = new FileRecordListResponse(responseFileRecords, total, skip, limit);
 		return response;
+	}
+
+	static mapScanResultParamsToDto(scanResultParams: ScanResultParams): ScanResultDto {
+		const scanResult = new ScanResultDto({
+			status: ScanStatus.VERIFIED,
+			reason: 'Clean',
+		});
+
+		if (scanResultParams.virus_detected) {
+			scanResult.status = ScanStatus.BLOCKED;
+			scanResult.reason = scanResultParams.virus_signature ?? 'Virus detected';
+		} else if (scanResultParams.error) {
+			scanResult.status = ScanStatus.WONT_CHECK;
+			scanResult.reason = scanResultParams.error;
+		} else if (scanResultParams.virus_detected === undefined) {
+			scanResult.status = ScanStatus.WONT_CHECK;
+			scanResult.reason = 'No scan result';
+		}
+
+		return scanResult;
 	}
 }
