+++ apps/server/src/modules/files-storage/mapper/index.ts
@@ -1,4 +1,4 @@
 export * from './copy-file-response.builder';
 export * from './file-dto.builder';
+export * from './file-record.mapper';
 export * from './files-storage.mapper';
-export * from './scan-result.mapper';
